requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.6.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
autopep8=2.0.4=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bleach=6.1.0=pypi_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
comm=0.2.0=pypi_0
cryptography=41.0.7=pypi_0
datetime=5.4=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
diskcache=5.6.3=pypi_0
docutils=0.20.1=pypi_0
fastjsonschema=2.19.0=pypi_0
fqdn=1.5.1=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.0=pypi_0
intervaltree=3.1.0=pypi_0
ipykernel=6.27.1=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jaraco-classes=3.3.0=pypi_0
jedi=0.19.1=pypi_0
jeepney=0.8.0=pypi_0
jinja2=3.1.2=pypi_0
json5=0.9.14=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.2=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.5.1=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.1=pypi_0
jupyter-server=2.12.1=pypi_0
jupyter-server-terminals=0.5.0=pypi_0
jupyterlab=4.0.9=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keyring=24.3.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
more-itertools=10.1.0=pypi_0
mpmath=1.3.0=pypi_0
nbclient=0.9.0=pypi_0
nbconvert=7.13.1=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
notebook=7.0.6=pypi_0
notebook-shim=0.2.3=pypi_0
openai=1.3.7=pypi_0
openssl=1.1.1w=h7f8727e_0
overrides=7.4.0=pypi_0
packaging=23.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pip=23.3.2=pypi_0
pkginfo=1.9.6=pypi_0
prometheus-client=0.19.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
psutil=5.9.7=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyautogen=0.2.2=pypi_0
pycodestyle=2.11.1=pypi_0
pycparser=2.21=pypi_0
pydantic=2.0.3=pypi_0
pydantic-core=2.3.0=pypi_0
python=3.9.0=hdb3f193_2
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
readme-renderer=42.0=pypi_0
referencing=0.32.0=pypi_0
requests=2.31.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986=2.0.0=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.15.2=pypi_0
scipy=1.11.4=pypi_0
secretstorage=3.3.3=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypi_0
terminado=0.18.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pypi_0
twine=4.0.2=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
tzdata=2023c=h04d1e81_0
uri-template=1.3.0=pypi_0
urllib3=2.1.0=pypi_0
wcwidth=0.2.12=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zope-interface=6.1=pypi_0
